Create calibrate-offsets.cfg
Initial commit of calibrate-offsets.cfg. This is stock from V's Tapchanger repo with the exception of commenting out [tools_calibrate] in its entirety, because I have no idea what PF5 is supposed to be connected to. I've heard it might be an external sensor. Even if it is, tool offsets can be calibrated with test prints with relative ease. Just including this here as a token inclusion since it's part of the original package.